+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/annotations/JsonIgnoreAnnotation.java
@@ -1,0 +1,10 @@
+package org.openprovenance.prov.template.compiler.past.annotations;
+
+public class JsonIgnoreAnnotation extends JavaAnnotation {
+
+    public static final String NAME = "java:@JsonIgnore";
+
+    public JsonIgnoreAnnotation() {
+        super(NAME);
+    }
+}